requirements.txt
streamlit==1.35.0
tensorflow
pandas==2.2.2
plotly-express==0.4.1
Pillow==10.3.0
seaborn==0.13.2
scikit-learn==1.5.0